Docs_Scrapping/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
python-docx
python-pptx
pandas
beautifulsoup4
lxml
pytesseract
pdf2image
Pillow
rouge-score
bert-score
jiwer
requests
